python-worker/requirements.txt
# Telegram Bot
python-telegram-bot==21.3

# Supabase
supabase==2.5.3

# AI - LLM & Transcription
groq==0.9.0
openai==1.35.3

# Azure Speech (TTS)
azure-cognitiveservices-speech==1.40.0

# Content Processing
youtube-transcript-api==0.6.2
httpx==0.27.0
beautifulsoup4==4.12.3
trafilatura==1.11.0
feedparser==6.0.11

# Audio Processing
pydub==0.25.1

# Web Server (for HTTP endpoint)
flask==3.0.3
gunicorn==23.0.0

# Utils
python-dotenv==1.0.1
structlog==24.2.0
